+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Only first Above The Fold audio post has a thumbnail and it's no longer cropped
+- Support form autovalues now set via Site Options: Fundraising Options metadata
 
 ## [3.7.1] - 2022-07-19